requirements.txt
# Django
django==1.10.7
django-bootstrap3==7.1.0

# Image processing
pillow==3.4.2
rawkit

# Data processing
numpy==1.11.2
pandas==0.19.1

# Threading
futures==3.0.5

# Utils
pyyaml
requests==2.12.3
